Revert "Minor database rebuild"

+++ src/logging.js
@@ -7,6 +7,14 @@ const fs = require('fs');
 const LOGGINGPATH = path.join("logs")
 const MAXLOGS = 10
 
+function parseDate(string) {
+    // 01234567
+    if (string.length !== 8) return null;
+    return new Date(
+        parseInt(string.slice(0, 4)), parseInt(string.slice(4, 6))-1, parseInt(string.slice(6))+1, 0, 0, 0, 0
+    ).getTime()
+}
+
 function stringifyDate(date) {
     const month = (date.getUTCMonth()+1).toString().padStart(2, '0')
     const day = date.getUTCDate().toString().padStart(2, '0')
@@ -22,7 +30,8 @@ async function removeOld() {
         .map(file => file.slice(0, -4))
         .sort((name1, name2) => name2 - name1)
 
-    while (logs.length > MAXLOGS) {
+    while (true) {
+        if (logs.length <= MAXLOGS) break;
         const oldFile = logs.pop()
         await fs.promises.rm(path.join(LOGGINGPATH, `${oldFile}.log`))
     }
